requirements.txt
aiogram
emoji
loguru
openai
Pillow
python-dotenv
requests
tenacity
